# improve performance for Matrix -> SparseMatrix