Benchmark (`sparse2` is PR):

```

A1 = rand(3000, 3000)
A2 = Matrix(sprand(3000, 3000, 0.1))
A3 = Matrix(sprand(3000, 3000, 0.01))

@btime sparse(A1)
  114.613 ms (11 allocations: 206.04 MiB)
@btime sparse(A2)
  27.586 ms (11 allocations: 20.61 MiB)
@btime sparse(A3)
  14.071 ms (11 allocations: 2.09 MiB)

@btime sparse2(A1)
  61.808 ms (7 allocations: 137.35 MiB)
@btime sparse2(A2)
  22.060 ms (7 allocations: 13.73 MiB)
@btime sparse2(A3)
  11.622 ms (7 allocations: 1.39 MiB
```