@mbauman 2018-03-06
Is there any reason why we wouldn't use this implementation for all `AbstractMatrix`es?  I suppose it'd probably be slower for sparse matrices… but most every other matrix implementation should be faster like this, no?