@KristofferC 2018-03-06
There are so many types of matrices I am not sure exactly what you need to fulfill the API. For example, this assumes column major and that indices starts at 1 etc. I was just playing it safe with `Matrix`.