@andreasnoack 2018-03-06
Extending to `StridedMatrix` should be safe though.